Adding tsi data.
You dont need to be changing any the load_previous or saving arguments to main(), just leave them as they are on GitHub (load_previous = TRUE, saving = TRUE). Inside data_and_reservations.R the beginning of the function says to generate the data if it doesnt exist for that climate model, otherwise it loads it. So if the data has been created before it and it exists on disk it will load it. And if it already exists it gives you a message that you wanted to save bc save = TRUE but that we will not save. See line 60:

if(load_previous){
    if(saving) message("You wanted saving but we wont be since we are loading in an already saved climate data set.")

Therefore, you dont need to change those arguments once you have the data.

Im currently running the code after deleting the previous slowtsi data from disk to recreate this data for you to run the experiment by sourcing the test_main.R file thats on Github from my last commit. I will report the results of this experiment and upload the generated slowtsi data when its done.