Made a few wordsmithing edits:
Added a direct link to the ODD at the top of the MODEL DESIGN section, and added a mention that the model code is also available at the project web site.

Fixed an erroneous description of risk.tak in the DISCUSSION section.

Minor wordsmithing to save a line here and there, where there was only one word in the last line of a paragraph, to keep the length to 12 pages.